+++ tests/unit/date-of-birth-or-age.directive.spec.js
@@ -3,7 +3,7 @@ describe('eha.date-of-birth-or-age.directive', function() {
   'use strict';
   beforeEach(function() {
     // Fake the current date so that our tests don't fail in the future
-    sinon.useFakeTimers(new Date(2015, 2, 1).getTime());
+    sinon.useFakeTimers(new Date(2015, 4, 3).getTime()); // Mai 3 2015
   });
 
   beforeEach(module('eha.date-of-birth-or-age.template'));
@@ -38,39 +38,38 @@ describe('eha.date-of-birth-or-age.directive', function() {
     yearsInput.dispatchEvent(new Event('input'));
 
     var monthsInput = $element[0].querySelector('input[name=ageMonths]');
-    monthsInput.value = 4;
+    monthsInput.value = 7;
     monthsInput.dispatchEvent(new Event('input'));
 
-    scope.$digest();
+    scope.$apply();
 
     var birthMonthSelect = $element[0].querySelector('select[name=birthMonth]');
     var birthYearSelect = $element[0].querySelector('select[name=birthYear]');
     var birthDaySelect = $element[0].querySelector('select[name=birthDay]');
 
     expect(birthYearSelect.value).to.eq('1982');
-    expect(birthMonthSelect.value).to.be.eq('10');
+    expect(birthMonthSelect.value).to.be.eq('9'); // 0 indexed
   });
 
   it('should calculate age from date of birth', function() {
 
-    // var birthDaySelect = $element[0].querySelector('select[name=birthDay]');
-    // birthDaySelect.value = '05';
-    // birthDaySelect.dispatchEvent(new Event('input'));
-    //
-    // var birthMonthSelect = $element[0].querySelector('select[name=birthMonth]');
-    // birthMonthSelect.value = '10';
-    // birthMonthSelect.dispatchEvent(new Event('input'));
-    //
-    // var birthYearSelect = $element[0].querySelector('select[name=birthYear]');
-    // birthYearSelect.value = '1982';
-    // birthYearSelect.dispatchEvent(new Event('input'));
+    var birthDaySelect = $element[0].querySelector('select[name=birthDay]');
+    birthDaySelect.value = '05';
+    birthDaySelect.dispatchEvent(new Event('change'));
 
-    scope.person.birthMonth = 10;
-    scope.person.birthYear = 1982;
+    var birthMonthSelect = $element[0].querySelector('select[name=birthMonth]');
+    birthMonthSelect.value = '9';
+    birthMonthSelect.dispatchEvent(new Event('change'));
+
+    var birthYearSelect = $element[0].querySelector('select[name=birthYear]');
+    birthYearSelect.value = '1982';
+    birthYearSelect.dispatchEvent(new Event('change'));
 
     scope.$digest();
 
-    var yearsInput = $element[0].querySelector('input[name=ageYears]');
-    var monthsInput = $element[0].querySelector('input[name=ageMonths]');
+    var ageYears = $element[0].querySelector('input[name=ageYears]');
+    var ageMonths = $element[0].querySelector('input[name=ageMonths]');
+    expect(parseInt(ageYears.value, 10)).to.equal(32);
+    expect(parseInt(ageMonths.value, 10)).to.equal(7);
   });
 });