+++ tests/unit/date-of-birth-or-age.service.spec.js
@@ -5,7 +5,7 @@ describe('eha.date-of-birth-or-age.service', function() {
   var service;
   beforeEach(module('eha.date-of-birth-or-age.service', function($provide) {
     // Fake the current date so that our tests don't fail in the future
-    sinon.useFakeTimers(new Date(2015, 2, 1).getTime());
+    sinon.useFakeTimers(new Date(2015, 4, 3).getTime()); //  Mai 3 2015
   }));
 
   beforeEach(inject(function(_ehaDateOfBirthOrAgeService_) {
@@ -16,7 +16,7 @@ describe('eha.date-of-birth-or-age.service', function() {
     describe('.getBirthDateFromAge()', function() {
       it('should return birth date when given an age', function() {
         var years = 32;
-        var months = 4;
+        var months = 7;
         var birthDate = service.getBirthDateFromAge(years, months);
 
         expect(birthDate.year).to.equal(1982);
@@ -31,7 +31,7 @@ describe('eha.date-of-birth-or-age.service', function() {
         var age = service.getAgeFromBirthDate(year, month);
 
         expect(age.years).to.equal(32);
-        expect(age.months).to.equal(4);
+        expect(age.months).to.equal(7);
       });
     });
   });